Extract all logic realted to events to separate class
