+++ src/plugin.ts
@@ -1,7 +1,8 @@
 import * as THREE from 'three';
-import type { Map as MapGL, MapPointerEvent } from '@2gis/mapgl/types';
+import type { Map as MapGL } from '@2gis/mapgl/types';
 
 import { Evented } from './external/evented';
+import { EventSource } from './eventSource';
 import { Loader } from './loader';
 import { PoiGroup } from './poiGroup';
 
@@ -34,13 +35,12 @@ export class GltfPlugin extends Evented<GltfPluginEventTable> {
     private camera = new THREE.PerspectiveCamera();
     private scene = new THREE.Scene();
     private tmpMatrix = new THREE.Matrix4();
-    private raycaster = new THREE.Raycaster();
-    private pointer = new THREE.Vector2();
     private viewport: ReturnType<HTMLElement['getBoundingClientRect']>;
     private map: MapGL;
     private options = defaultOptions;
     private loader: Loader;
     private poiGroup: PoiGroup;
+    private eventSource?: EventSource;
     private onPluginInit = () => {}; // resolve of waitForPluginInit
     private waitForPluginInit = new Promise<void>((resolve) => (this.onPluginInit = resolve));
     private models;
@@ -89,7 +89,7 @@ export class GltfPlugin extends Evented<GltfPluginEventTable> {
         map.once('idle', () => {
             this.poiGroup.addIcons();
             this.addThreeJsLayer();
-            this.bindEvents();
+            this.initEventHandlers();
         });
     }
 
@@ -169,93 +169,20 @@ export class GltfPlugin extends Evented<GltfPluginEventTable> {
     private invalidateViewport() {
         const container = this.map.getContainer();
         this.viewport = container.getBoundingClientRect();
+        this.eventSource?.updateViewport(this.viewport);
     }
 
-    private getEventTargetMesh(e: MouseEvent | TouchEvent) {
-        const { clientX, clientY } = 'changedTouches' in e ? e.changedTouches[0] : e;
-
-        // coordinates of the cursor in local coordinates of map's viewport
-        const localX = clientX - this.viewport.x;
-        const localY = this.viewport.height - (clientY - this.viewport.y);
-
-        // convert local coordinates of the corser to WebGL coordinates
-        // and use it for the object identification in three.js-scene
-        this.pointer.x = (localX / this.viewport.width) * 2 - 1;
-        this.pointer.y = (localY / this.viewport.height) * 2 - 1;
-        this.raycaster.setFromCamera(this.pointer, this.camera);
-        const intersects = this.raycaster.intersectObjects(this.scene.children, true);
-        const target = intersects[0] ? intersects[0] : undefined;
-
-        if (!target || target.object.type !== 'Mesh') {
-            return null;
-        }
-
-        return target;
-    }
-
-    private isGeoJsonPoi(e: MapPointerEvent) {
-        return (
-            e.targetData?.type === 'geojson' &&
-            e.targetData?.feature?.properties?.type === 'immersive_poi'
-        );
-    }
-
-    private bindEvents() {
+    private initEventHandlers() {
         this.map.on('resize', () => {
             this.invalidateViewport();
         });
 
-        /**
-         * Poi events
-         */
-        this.map.on('mousemove', (e) => {
-            if (this.isGeoJsonPoi(e)) {
-                this.emit('mousemovePoi', e);
-            }
-        });
-
-        this.map.on('mouseover', (e) => {
-            if (this.isGeoJsonPoi(e)) {
-                this.emit('mouseoverPoi', e);
-            }
-        });
-
-        this.map.on('mouseout', (e) => {
-            if (this.isGeoJsonPoi(e)) {
-                this.emit('mouseoutPoi', e);
-            }
-        });
-
-        this.map.on('click', (e) => {
-            if (this.isGeoJsonPoi(e)) {
-                this.emit('clickPoi', e);
-            }
-        });
-
-        /**
-         * Model events
-         */
-        this.map.on('click', (e) => {
-            const target = this.getEventTargetMesh(e.originalEvent);
-
-            if (target) {
-                this.emit('clickModel', {
-                    lngLat: e.lngLat,
-                    point: e.point,
-                    originalEvent: e.originalEvent,
-                    target: {
-                        id: target.object.userData.modelId,
-                    },
-                });
-            }
-        });
-
-        this.map.on('mousemove', (e) => {
-            const target = this.getEventTargetMesh(e.originalEvent);
-            if (target) {
-                console.log('hover over the model', Date.now());
-            }
-        });
+        this.eventSource = new EventSource(this.map, this.viewport, this.camera, this.scene);
+        for (let eventName of this.eventSource.getEvents()) {
+            this.eventSource.on(eventName, (e) => {
+                this.emit(eventName, e);
+            });
+        }
     }
 
     private render() {
