+++ src/types/utils.ts
@@ -1,1 +1,3 @@
 export type Parameter<T extends (...args: any) => any> = Parameters<T>['0'];
+
+export type ViewportDom = ReturnType<HTMLElement['getBoundingClientRect']>;
